% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmInit.R
\docType{class}
\name{AmInit-class}
\alias{AmInit-class}
\title{AMARETTO initialization data structure}
\description{
AMARETTO initialization data structure
}
\section{Slots}{

\describe{
\item{\code{MA_matrix_Var}}{describe here}

\item{\code{RegulatorData}}{describe here}

\item{\code{RegulatorAlterations}}{describe here}

\item{\code{ModuleMembership}}{describe here}

\item{\code{Parameters}}{describe here}

\item{\code{NrCores}}{describe here -- should it be part of class?}
}}

