% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmInit.R
\name{AmInit}
\alias{AmInit}
\title{constructor for AmInit instances, built from list}
\usage{
AmInit(initlist)
}
\arguments{
\item{initlist}{a list with obligatory elements}
}
\description{
constructor for AmInit instances, built from list
}
\examples{
ini = AmInit(AmObjects::LIHC_AMARETTOinit)
ini
}
