% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AmInit.R
\name{AmResults}
\alias{AmResults}
\title{constructor for AMARETTO results object}
\usage{
AmResults(reslist)
}
\arguments{
\item{reslist}{list with obligatory components}
}
\description{
constructor for AMARETTO results object
}
\examples{
res = AmObjects::LIHC_AMARETTOresults
AmResults(res)
}
